Determine the current index for the TypeSense container
This change allows the operator to override the statefulset template structure in our own operator image by including other sidecar containers that typesense requires to export metrics for Prometheus, backups to GCS or S3, etc.

Using the dynamic index the operator will be able to determine what the index of the typesense container is within the pod and update its specific parameters.